SearchFrag.java 내 장소 리스트 출력 Volley -> Retrofit2로 교체. BaseFrag.java 상속 의미 X
