+++ app/src/main/java/com/example/coditplace2/BaseFrag.java
@@ -35,8 +35,6 @@ import retrofit2.Callback;
 public class BaseFrag extends Fragment {
     //Volley
     Map<String, String> params = new HashMap<String, String>();
-    //Retrofit
-    HashMap<String, String> params2 = new HashMap<>();
 
     public BaseFrag() {
     }
@@ -70,34 +68,4 @@ public class BaseFrag extends Fragment {
             Log.d("ErrorResponse", "통신 실패");
         }
     };
-    //Retrofit
-    public void requestR(String type, String search, ArrayList<ItemData> arr){
-        //검색어 넣어주기
-        params2.put(type, search);
-        RetroClient.getRetroBaseApiService().rPlocation(params2).enqueue(new Callback<List<ResponseGet>> () {
-            @Override
-            public void onResponse(Call<List<ResponseGet>>  call, retrofit2.Response<List<ResponseGet>>  response) {
-                List<ResponseGet> result = response.body();
-                Log.d("retrofit", "onResponse: success1" + result); //200 정상통신
-                Log.d("retrofit", "onResponse: success2"+ result.get(0)+"/"+result.get(1));
-                for(int i = 0; i<result.size(); i++){
-                    String pidx = String.valueOf(result.get(i).getPidx());
-                    String pname = result.get(i).getPname();
-                    String pimage1 = result.get(i).getPimage1();
-                    String pvisit = result.get(i).getPvisit();
-                    String picon = (String) result.get(i).getPicon();
-                    String pcategory = String.valueOf(result.get(i).getPcategory());
-                    String pphone = result.get(i).getPphone();
-                    String pcontent = result.get(i).getPcontent();
-
-                    arr.add(i, new ItemData(pidx, pname, pimage1, pvisit, picon, pcategory, pphone, pcontent, "0"));
-                    Log.d("retrofit", "arr:" + arr.get(i).getpName());
-                }
-            }
-            @Override
-            public void onFailure(Call<List<ResponseGet>>  call, Throwable t) {
-                Log.d("retrofit", "onResponse: failed");
-            }
-        });
-    }
 }
